pyproject.toml
# pyproject.toml(Py 3.9 + NumPy 2.0.2) 기준

# =============================================
# 빌드 시스템 설정 (PEP 518)
# =============================================
[build-system]
# 빌드에 필요한 패키지 목록 (Cython과 NumPy 포함, NumPy 2.x 헤더 사용)
requires = [
    "setuptools>=61",
    "wheel",
    "cython>=0.29", # Cython 버전은 필요시 조정
    "numpy>=2.0",   # NumPy 2.x 헤더 사용 (NumPy 2.x 버전 사용), 기존 numpy>=2.1
]
build-backend = "setuptools.build_meta"
#backend-path = ["."] # setup.py(ext_modules 정의용)를 사용하기 위해, 일단 불필요


# =============================================
# 프로젝트 메타데이터 (PEP 621)
# =============================================
[project]
name = "bok-da"                 
version = "0.3.1"               
authors = [
  {name = "Data Science Team, Bank of Korea"},
  {email = "di@bok.or.kr" },
]
maintainers = [
  { name="Changhoon Lee", email="changhoon@bok.or.kr"}
]
description = "BOK Python Library for Generalized Data Analysis"
readme = "README.md"
requires-python = ">=3.9"      # ★ 목표 환경 반영, 외부망에선 >=3.11도 호환
license = { file="LICENSE.txt" }    # LICENSE 파일이 루트에 있다고 가정
keywords = ["bok", "data analysis", "econometrics", "time series", "Bayesian"]
classifiers = [                 
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: BSD 3-Clauses License", # 라이선스는 추후 확인 필요
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Topic :: Economics",
    "Topic :: Ecnomometrics",
    "Topic :: Time Series Analysis",
    "Topic :: Bayesian Inference", ## 추후 수정 필요
]

# =============================================
# 런타임 의존성 (Py3.9 + NumPy 2.0.2 기준 통합 및 조정)
# =============================================
dependencies = [
    "numpy>=1.26,<2.2",        # NumPy 2.x 버전 사용 (NumPy 2.x 헤더 사용), numpy>=2.1,<2.2
    "pandas>=2.0",            
    "scipy>=1.13,<1.15",       # 2차 setup.py 기준 (Py3.11/NumPy2.x 호환 확인됨), 기존 scipy>=1.14,<1.15
    "statsmodels>=0.14,<0.15",
    "matplotlib>=3.8.2",      
    "seaborn>=0.13.0",        
    "linearmodels>=6.0",      # linearmodels>=6.0
    "openpyxl",               
    "tqdm",                   
    "threadpoolctl>=3.5,<3.6", # 2차 setup.py 기준
    #"wrapt>=1.14",
    #"mypy_extensions>=1.0",
    "requests>=2.28,<3.0",        # 2차 setup.py 기준
    # "s3fs>=2023.12.2",        # 내부망용
    # "cython"은 빌드 의존성이므로 여기서 제외
    # 'wrapt', 'mypy_extensions'는 다른 패키지의 하위 의존성이므로 직접 명시 불필요
]

# =============================================
# 선택적 의존성 (개발/테스트용)
# =============================================
[project.optional-dependencies]
test = [
    "pytest>=7.0",
    "pytest-cov",
]
dev = [
    "bok-da[test]",
    "ruff",
    "build",
    "twine",
    "sphinx",
    "jupyterlab",
    "notebook",
    "ipykernel",
]

# =============================================
# 프로젝트 관련 URL (추후 수정 필요)
# =============================================
[project.urls]
Homepage = "https://bidas-gitlab.boknet.intra/digitaltech"
Repository = "https://bidas-gitlab.boknet.intra/digitaltech/bok_library.git"

# =============================================
# Setuptools 관련 추가 설정
# =============================================
[tool.setuptools]
zip-safe = false # C 확장 모듈 포함으로 인한 zip-safe=False 설정

[tool.setuptools.packages.find]
where = ["src"]       # 소스 코드는 'src' 디렉토리 아래에 위치
include = ["bok_da*"] # 'bok_da' 로 시작하는 모든 패키지 포함

## --- (참고) 패키지 내 데이터 파일 포함 ---
## [tool.setuptools.package-data]
## bok_da = ["data_files/*.csv"]